pyproject.toml
[project]
name = "github-token-exchange"
version = "0.1.0"
description = "Secure token exchange service: Convert GitHub Actions OIDC tokens to GitHub Application installation access tokens"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pre-commit>=3.7.1",
    "ruff>=0.4.4",
    "pytest-cov>=6.1.1",
    "mypy>=1.16.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff.lint]
preview = true
# "I" = imports are sorted (following isort rules above)
# "PLC0415" = all imports must be at top of the file, imports within functions are not allowed
extend-select = ["I", "PLC0415"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
incremental = false

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing"
pythonpath = ["."]
